pyproject.toml
[project]
name = "aws-cost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-lambda-typing>=2.20.0",
    "dotenv>=0.9.9",
    "google-genai>=1.50.1",
    "mcp>=1.21.2",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "strands-agents>=1.18.0",
    "strands-agents-tools>=0.2.16",
]
